move to opened_video instance
