+++ packages/renderer/rust/main.rs
@@ -4,6 +4,7 @@ mod errors;
 mod ffmpeg;
 mod global_printer;
 mod image;
+mod opened_video;
 mod payloads;
 use commands::execute_command;
 use errors::PossibleErrors;
